my_telegram_bot/requirements.txt
python-telegram-bot
pillow==10.1.0
numpy==1.26.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.8.2
pandas==2.1.3 